infrastructure/cdk.out/asset.b8f91edf3200d7dc09935b4b7006cd871a7c691461906873084af8d3ba97bd70/core/schemas.py
"""
Schemas for data validation in the Songs API.
"""

from marshmallow import Schema, fields, validate

class SongSchema(Schema):
    """Schema for validating song data."""
    song_id = fields.String(dump_only=True)  # Read-only field
    title = fields.String(required=True, validate=validate.Length(min=1))
    artist = fields.String(required=True, validate=validate.Length(min=1))
    lyrics = fields.String(required=True, validate=validate.Length(min=1))

# Create instances for reuse
song_schema = SongSchema()
songs_schema = SongSchema(many=True) 