infrastructure/requirements.txt
# Infrastructure dependencies (CDK)
aws-cdk-lib>=2.0.0
constructs>=10.0.0

# Application dependencies
boto3==1.34.34
python-dotenv==1.0.1 